Add ROI area calculation and export to ROI list
This update adds calculation of ROI area in square microns (or pixels if pixel size is unavailable) to the ROI export in roi_list.py. Pixel size is extracted from metadata if available, and the area is included as a new column in the exported data. Minor cleanup was also performed in meta_info.py.